Burger Flippin'
Assume n customers all arrive and place an order. The orders are close enough to simultaneous that you may consider them all to be at the same time. You now need to flip their burgers, put on condiments, and arrange the order but you can only work on one order at a time. Each order may take a varying amount of time, so you will have to work intelligently to pick which order to work on first, second, and so on.

Solution: 
In order to have the shortest wait time for each customer, you much complete the orders that take the least time first.
Output: the average amount of time that each customer waited.