Profile Picture Scaling
Input:
Input will consist of a single test case. The input will be on a single line, and will contain 3 integer numbers, N (N>0), W (W>0), and H (H>0). N represents the size of the square box in which all profile pictures must fit. W and H represent the width and height of the profile picture that needs to be resized.

Output:
Output a decimal number of how much the profile picture needs to be scaled such that the maximum of the new height and width is equal to the size of the square the picture will sit in. Output should be rounded to four decimal places.