Speeding Advantage
Input:
There will be up to 10 test cases in the input. Each test case will consist of three integers on a single line d (1 < d < 1000), s1 and s2 (1 < s1 < s2 < 1000), where d is a distance in miles, and s1 and s2 are speeds in mph. The integers will be separated by a single space and there will be no leading or trailing blanks. The input will end with a line of three 0s.

Output:
For each case, output a time in the form of H:MM:SS, representing the difference in time to travel the distance.